+++ lib/components/features/event-registration/modals/RegistrationModal.tsx
@@ -1,6 +1,6 @@
-import { intersection, partition, uniqBy } from "lodash";
+import { intersection, partition } from "lodash";
 import { useMe } from "$lib/hooks/useMe";
-import { userAvatar } from "$lib/utils/user";
+import { communityAvatar } from "$lib/utils/community";
 import { useSession } from "$lib/hooks/useSession";
 
 import { Avatar, Button, Segment } from "$lib/components/core";
@@ -80,7 +80,7 @@ export function RegistrationModal() {
             {
               me && (
                 <div className='flex gap-3'>
-                  <Avatar src={userAvatar(me)} size="xl" />
+                  <Avatar src={communityAvatar(me)} size="xl" />
                   <div>
                     <p className='font-medium'>{me.name}</p>
                     <p className='text-secondary text-[14px]'>{me.email}</p>