+++ lib/components/features/event/EventLocationBlock.tsx
@@ -6,13 +6,12 @@ import { isAttending } from '$lib/utils/event';
 export function EventLocationBlock({ loading = false, event }: { loading?: boolean; event?: Event }) {
   const session = useSession();
 
+  const attending = session?.user && event ? isAttending(event, session?.user) : false;
+  
   if (loading) return <EventLocationBlockSkeleton />;
-  if (!event?.address) return null;
 
-  const attending = session?.user ? isAttending(event, session?.user) : false;
-
-  return (
-    <div className="flex gap-4 flex-1">
+  if (event?.address) return (
+    <div className="flex gap-4 flex-1 w-full">
       <div className="border rounded-sm size-12 min-w-12 flex items-center justify-center">
         <i className="icon-location-outline" />
       </div>
@@ -32,6 +31,17 @@ export function EventLocationBlock({ loading = false, event }: { loading?: boole
       </div>
     </div>
   );
+
+  if (event?.virtual_url) return (
+    <div className="flex gap-4 flex-1 items-center">
+      <div className="border rounded-sm size-12 min-w-12 flex items-center justify-center">
+        <i className="icon-video" />
+      </div>
+      <p>Virtual</p>
+    </div>
+  );
+
+  return null;
 }
 
 function EventLocationBlockSkeleton() {